Move model.ts to src/db
